requirements.txt
# Cookidoo Agent Assistant - Combined Requirements (Local Development)
# Python 3.11+
#
# This file combines backend + mcp-server dependencies for a single
# local development venv. The individual requirements files in backend/
# and mcp-server/ are kept for Docker container builds.

# ─── FastAPI and server ────────────────────────────────────────────
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ─── Databricks and MLflow ─────────────────────────────────────────
databricks-sdk>=0.14.0
databricks-agents>=0.2.0
mlflow>=2.9.2

# ─── Azure integrations ───────────────────────────────────────────
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0

# ─── LLM and AI ───────────────────────────────────────────────────
anthropic>=0.7.7
openai>=1.3.7
langchain>=0.1.0
langchain-community>=0.0.1

# ─── Data processing ──────────────────────────────────────────────
pandas>=2.1.3
numpy>=1.26.2
pyarrow>=14.0.1

# ─── HTTP clients ─────────────────────────────────────────────────
httpx>=0.25.2
requests>=2.31.0
aiohttp>=3.9.1

# ─── Async support ────────────────────────────────────────────────
aioboto3>=12.1.0

# ─── Utilities ────────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ─── Logging and monitoring ───────────────────────────────────────
structlog>=23.2.0
python-json-logger>=2.0.7
sentry-sdk>=1.38.0

# ─── MCP Server (Cookidoo API + MCP protocol) ─────────────────────
cookidoo-api>=0.5.0
fastmcp>=0.1.0
mcp>=1.0.0

# ─── Testing ──────────────────────────────────────────────────────
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
faker>=20.1.0

# ─── Code quality ─────────────────────────────────────────────────
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
isort>=5.12.0
pre-commit>=3.5.0

# ─── Documentation ────────────────────────────────────────────────
mkdocs>=1.5.3
mkdocs-material>=9.4.14
